Enhanced ProviderDetection analysis and added new tests when analysis generated RuntimeExceptions
Firstly, a String list of all default CryptSL rules is loaded from the rules dir, and not created a whitelist like before.
Secondly, now the analysis chooses all the rules from detected provider, plus default rules that are not still implemented for detected provider.
Lastly, added new JUnit tests for when RuntimeExceptions are thrown, in the case when provider var flows through IF-ELSE, SWITCH stmt or TERNARY operators.