Fixes eclipse-cognicrypt/CogniCrypt#293. Re-compiling of rules required.
Introducing a 'hop' mechanic to state machines that shows how many steps
are required to reach an accepting state. May be used to calculate path
to successful operation. The fix uses it to show only alternative calls
if they are in fact closer to an accepting state than the current state.